add application exception

+++ migrations/migration_00104.php
@@ -56,7 +56,7 @@ try
     `sci_id` int(11) DEFAULT NULL,
     
         
-    applicant_id int(11) NOT NULL , 
+    applicant_id bigint(20) NOT NULL , 
     application_plan_id int(11) NOT NULL , 
     application_simulation_id int(11) NOT NULL , 
     expiry_date varchar(8) NOT NULL ,
